added include <time.h> to log.cpp and fixed some signed <-> unsigned warnings
git-svn-id: svn://maxr.org/MAX/trunk@2313 c3bd5e88-ca0b-48cf-b972-da51c77c4c1b

+++ menus.cpp
@@ -2751,7 +2751,7 @@ cNetworkHostMenu::~cNetworkHostMenu()
 	delete loadButton;
 	delete startButton;
 	
-	for (int i = 0; i < mapSenders.size (); i++)
+	for (unsigned int i = 0; i < mapSenders.size (); i++)
 	{
 		if (mapSenders[i] != 0)
 			delete mapSenders[i];
@@ -2984,7 +2984,7 @@ void cNetworkHostMenu::handleNetMessage( cNetMessage *message )
 			if (gameDataContainer.map != 0 && MapDownload::isMapOriginal (gameDataContainer.map->MapName) == false)
 			{
 				int receiverNr = message->popInt16 ();
-				if (receiverNr >= 0 && receiverNr < players.Size ())
+				if (receiverNr >= 0 && receiverNr < (int)players.Size ())
 				{
 					int socketNr = players[receiverNr]->socket;
 					cMapSender* mapSender = new cMapSender (socketNr, gameDataContainer.map->MapName);
